feat: update user login

+++ tools/generate_secret.go
@@ -1,4 +1,4 @@
-package main
+package secret
 
 import (
 	"crypto/rand"
@@ -7,6 +7,7 @@ import (
 	"io/ioutil"
 	"log"
 	"os"
+	"strings"
 )
 
 func generateRandomSecret(length int) (string, error) {
@@ -19,19 +20,30 @@ func generateRandomSecret(length int) (string, error) {
 	return base64.URLEncoding.EncodeToString(randomBytes), nil
 }
 
-func main() {
-	secret, err := generateRandomSecret(32)
-	if err != nil {
-		log.Fatal("Error generating secret:", err)
-	}
-
+func EnsureJwtSecret() {
 	envFile := ".env"
-	envContent := fmt.Sprintf("JWT_SECRET=%s\n", secret)
 
-	err = ioutil.WriteFile(envFile, []byte(envContent), os.FileMode(0600))
+	// Read the existing content of the .env file
+	existingContent, err := ioutil.ReadFile(envFile)
 	if err != nil {
-		log.Fatal("Error writing secret to .env file:", err)
+		log.Fatal("Error reading .env file:", err)
 	}
 
-	fmt.Println("JWT Secret generated and stored in .env file.")
+	if !strings.Contains(string(existingContent), "JWT_SECRET=") {
+		secret, err := generateRandomSecret(32)
+		if err != nil {
+			log.Fatal("Error generating secret:", err)
+		}
+
+		// Append the JWT secret to the existing content
+		envContent := fmt.Sprintf("%sJWT_SECRET=%s\n", string(existingContent), secret)
+
+		// Write the new content back to the .env file
+		err = ioutil.WriteFile(envFile, []byte(envContent), os.FileMode(0600))
+		if err != nil {
+			log.Fatal("Error writing secret to .env file:", err)
+		}
+
+		fmt.Println("JWT Secret generated and stored in .env file.")
+	}
 }
